refactor: centralize boto3 availability checks in conftest.py for cleaner test files
