+++ libs/langgraph-checkpoint-aws/tests/integration_tests/store/dynamodb/test_dynamodb_store_integration.py
@@ -11,12 +11,7 @@ import pytest
 from langgraph_checkpoint_aws.store.dynamodb import DynamoDBStore
 from langgraph_checkpoint_aws.store.dynamodb.exceptions import ValidationError
 
-try:
-    import boto3
-
-    BOTO3_AVAILABLE = True
-except ImportError:
-    BOTO3_AVAILABLE = False
+from .conftest import BOTO3_AVAILABLE, DYNAMODB_AVAILABLE
 
 pytestmark = pytest.mark.skipif(
     not BOTO3_AVAILABLE,
@@ -25,39 +20,6 @@ pytestmark = pytest.mark.skipif(
 
 if BOTO3_AVAILABLE:
     from langgraph.store.base import Item, SearchItem
-
-
-def _is_dynamodb_available() -> bool:
-    """Check if a DynamoDB Local server is available for testing.
-
-    Verifies availability by making a real DynamoDB API call (list_tables)
-    with dummy credentials, rather than just checking if the port is open.
-    """
-    if not BOTO3_AVAILABLE:
-        return False
-
-    from .conftest import (
-        DYNAMODB_AWS_ACCESS_KEY_ID,
-        DYNAMODB_AWS_SECRET_ACCESS_KEY,
-        DYNAMODB_ENDPOINT_URL,
-        DYNAMODB_REGION,
-    )
-
-    try:
-        client = boto3.client(
-            "dynamodb",
-            region_name=DYNAMODB_REGION,
-            endpoint_url=DYNAMODB_ENDPOINT_URL,
-            aws_access_key_id=DYNAMODB_AWS_ACCESS_KEY_ID,
-            aws_secret_access_key=DYNAMODB_AWS_SECRET_ACCESS_KEY,
-        )
-        client.list_tables(Limit=1)
-        return True
-    except Exception:
-        return False
-
-
-DYNAMODB_AVAILABLE = _is_dynamodb_available()
 
 
 @pytest.fixture
